deployment/requirements.txt
boto3==1.26.69
PyYAML==6.0
